InvestoChat_Build/requirements.txt
requests>=2.32.0
python-dotenv>=1.0.1
psycopg[binary]>=3.1.18
openai>=1.58.1
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pymupdf>=1.24.10
tenacity>=8.2.3